requirements.txt
fastapi
uvicorn
gunicorn
google-cloud-firestore
python-multipart